+++ crates/lkjscript-compiler/src/analyze/declarations/products.rs
@@ -6,7 +6,7 @@ impl Analyzer {
         program: &ValidatedSourceTree,
     ) -> Result<()> {
         for (source_index, file) in program.files().iter().enumerate() {
-            let source_raw = u32::try_from(source_index)
+            let source_raw = u64::try_from(source_index)
                 .map_err(|_| Error::msg("too many source files for HIR SourceId"))?;
             let source = SourceId::new(source_raw);
             for form in &file.forms {
