+++ crates/lkjscript-compiler/src/hir/memory_plan/producer.rs
@@ -21,8 +21,8 @@ struct Producer<'a> {
     entries: Vec<MemoryPlanEntry>,
     expression_entries: HashMap<MemoryExpressionId, MemoryEntryId>,
     child_entries:
-        HashMap<MemoryExpressionId, Vec<(u32, MemoryExpressionId, Option<MemoryDropPathId>)>>,
-    places_by_binding: HashMap<(MemoryFunctionId, u32), u32>,
+        HashMap<MemoryExpressionId, Vec<(u64, MemoryExpressionId, Option<MemoryDropPathId>)>>,
+    places_by_binding: HashMap<(MemoryFunctionId, u64), u64>,
     products_by_id: HashMap<hir::ProductId, usize>,
     enums_by_id: HashMap<hir::EnumId, usize>,
     uses: Vec<MemoryUse>,
@@ -33,8 +33,8 @@ struct Producer<'a> {
     destinations: Vec<MemoryDestinationPlan>,
     borrow_scopes: Vec<MemoryBorrowScopePlan>,
     current_function: MemoryFunctionId,
-    next_expression: u32,
-    next_place: u32,
+    next_expression: u64,
+    next_place: u64,
     expression_parents: BTreeMap<MemoryExpressionId, Option<MemoryExpressionId>>,
     work: MemoryPlanWork,
 }
